+++ src/main/java/com/ulegalize/utils/PrestationUtils.java
@@ -19,7 +19,7 @@ public class PrestationUtils {
      * @param vat         the vat
      * @return the big decimal
      */
-    public static BigDecimal calculateVAT(boolean forfait, BigDecimal minutes, BigDecimal hours, Integer couthoraire, BigDecimal forfaitHt, BigDecimal vat) {
+    public static BigDecimal calculateVAT(boolean forfait, BigDecimal minutes, BigDecimal hours, BigDecimal couthoraire, BigDecimal forfaitHt, BigDecimal vat) {
 
         if (!forfait) {
             // ( (( (dh * 60) + dm ) / 60 ) * couthoraire ) * (1 + (vat / 100))
@@ -40,12 +40,12 @@ public class PrestationUtils {
      * @param forfaitHt   the forfait ht
      * @return the big decimal
      */
-    public static BigDecimal calculateHVAT(boolean forfait, BigDecimal minutes, BigDecimal hours, Integer couthoraire, BigDecimal forfaitHt) {
+    public static BigDecimal calculateHVAT(boolean forfait, BigDecimal minutes, BigDecimal hours, BigDecimal couthoraire, BigDecimal forfaitHt) {
         if (!forfait) {
             // (( (dh * 60) + dm ) / 60 ) * couthoraire
             BigDecimal result = ((hours.multiply(BigDecimal.valueOf(60))).add(minutes))
                     .divide(BigDecimal.valueOf(60L), 6, RoundingMode.HALF_EVEN)
-                    .multiply(BigDecimal.valueOf(couthoraire));
+                    .multiply(couthoraire);
 
 // Set scale to 2 decimal places with HALF_UP rounding mode
             result = result.setScale(2, RoundingMode.HALF_UP);
